Added special handling for MS token length of `2`.
Fixes #54

There is no need to handle the case of MS token length of 3, since that will naturally format as expected for timers.
It seems like a timer with a single MS digit would be unlikely -- and it would be problematic at best to handle a forced length of `1` gracefully and intuitively based on the behaviour for every other token.